Changes supervisor config setup
The new setup uses the Debian supervisor package's shipped config, with the additions we need for Apache2 installed as an included file in `/etc/supervisor/conf.d/`.

This replaces the single Supervisor config we were installing, which had a mix of Supervisor v2 and v3 compatible config. This was causing `supervisorctl` to complain it couldn't find  the configured socketfile used for communicating with the Supervisor service.